[CodeQuality][#74] Fix possible loss of precision warnings (C4244)

+++ Tympan/geometric_methods/AcousticRaytracer/Acoustic/Diffraction.cpp
@@ -57,13 +57,13 @@ void getThetaRandom(const decimal& angleOuverture, const decimal & delta_theta, 
 
 	if (theta > angleOuverture / 2.)
 	{
-		theta += (2 * M_PI - angleOuverture);
+		theta += ((decimal)(2 * M_PI) - angleOuverture);
 	}
 }
 
 void getThetaRegular(const decimal& angleOuverture, const decimal & delta_theta, const decimal & nbResponseLeft, decimal& theta)
 {
-	theta = (nbResponseLeft * delta_theta) - (angleOuverture / 2.);
+	theta = (nbResponseLeft * delta_theta) - (angleOuverture / 2.f);
 }
 
 
@@ -124,7 +124,7 @@ bool Diffraction::getResponse(vec3& r, bool force)
 	do
 	{
 		decimal theta = 0.;
-		(*getTheta) (angleOuverture, delta_theta, nbResponseLeft, theta);
+		(*getTheta) (angleOuverture, delta_theta, (decimal)nbResponseLeft, theta);
 
 #ifdef _ALLOW_TARGETING_
 		if (!force)
@@ -205,7 +205,7 @@ bool Diffraction::generateTest(std::vector<vec3>& succededTest, std::vector<vec3
 {
     for (unsigned int i = 0; i < nbResponses; i++)
     {
-        vec3 newDir = vec3(((decimal)rand() / (decimal)RAND_MAX) * 2. - 1., ((decimal)rand() / (decimal)RAND_MAX) * 2. - 1., ((decimal)rand() / (decimal)RAND_MAX) * 2. - 1.);
+        vec3 newDir = vec3(((decimal)rand() / (decimal)RAND_MAX) * 2.f - 1.f, ((decimal)rand() / (decimal)RAND_MAX) * 2.f - 1.f, ((decimal)rand() / (decimal)RAND_MAX) * 2.f - 1.f);
         newDir.normalize();
         if (isAcceptableResponse(newDir))
         {
